backend/requirements.txt
tensorflow-cpu==2.15.0
keras>=3.0.0
fastapi
uvicorn
python-multipart
numpy<2.0.0
Pillow
opencv-python
shap
matplotlib
requests
